Importing Experiment Parameters

Importing experiment parameters from a static file allows for batch processing of experiments that may have asymmetries, e.g. drug, stimulus, experiment id. User should manually fill in a "experiment_parameters.csv" file. The function, create_parameter_template, can be called to generate a template file.

Usage

import_experiment_parameters(file = NULL)

Arguments

file

Data.Frame or Character (PATH to .csv file). Users can pass a data frame or tibble into this function or the full PATH to a .csv file. This allows for the use of a remote/cloud-based file, e.g. googlesheets, or local file. Default is NULL and will prompt the user to interactively select the file.

Value

Environment objects (variables) to matools_env:

parameters

experiment_types

See also

Examples

if (FALSE) {
library(matools)

### Passing the PATH to a local file ###
local_file_path <- "/path/to/my/parameters.csv"
import_experiment_parameters(file = local_file_path)

### Using a publicly shared Google Sheet ###
googlesheets4::gs4_deauth()
public_sheet <- "https://docs.google.com/spreadsheets/d/YOUR_SHEET_HASH_HERE/edit?usp=sharing"

# Convert the sheet to a data frame
parameters <-
  googlesheets4::read_sheet(public_sheet)
import_experiment_parameters(file = parameters)
}